// src/utils/types.ts
export type SubCategory = {
  product_id: number
  title: string
  image: string
  icon?: string
  subcategories: InnerSubCategory[]
}

export type InnerSubCategory = {
  product_id: number
  title: string
  image?: string
}

export type Category = {
  product_id: number
  title: string
  image: string
  color: string
  icon?: string
  subcategories: SubCategory[]
}